requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
Flask-Dance==7.0.0
Flask-Cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
